requirements.txt
flask
paho-mqtt
numpy
opencv-python-headless
insightface
onnxruntime
faiss-cpu
pytest
python-dotenv
